1st version of NRPPa Position Information Transfer Procedure

+++ openair3/NGAP/ngap_gNB_itti_messaging.c
@@ -80,7 +80,7 @@ void ngap_gNB_itti_send_sctp_close_association(instance_t instance, int32_t asso
 }
 
 
-void ngap_gNB_itti_send_DownlinkUEAssociatedNRPPa(instance_t instance, uint32_t gNB_ue_ngap_id, uint32_t amf_ue_ngap_id,  uint8_t *routingId_buffer, uint32_t routingId_buffer_length, uint8_t *nrppa_pdu, uint32_t nrppa_pdu_length){
+void ngap_gNB_itti_send_DownlinkUEAssociatedNRPPa(instance_t instance, uint32_t gNB_ue_ngap_id, uint64_t amf_ue_ngap_id,  uint8_t *routingId_buffer, uint32_t routingId_buffer_length, uint8_t *nrppa_pdu, uint32_t nrppa_pdu_length){
 
   MessageDef          *message_p;
   ngap_DownlinkUEAssociatedNRPPa_t *ngap_DownlinkUEAssociatedNRPPa;
